ajustando controller
