+++ src/main/kotlin/br/com/jhonatansouza/starbuckets/service/ProviderService.kt
@@ -1,4 +1,4 @@
-package br.com.jhonatansouza.starbuckets.service.impl
+package br.com.jhonatansouza.starbuckets.service
 import br.com.jhonatansouza.starbuckets.exception.GenericException
 import br.com.jhonatansouza.starbuckets.model.Provider
 import br.com.jhonatansouza.starbuckets.model.request.ProviderRequest
